+++ internal/repos/home.go
@@ -1,38 +1,26 @@
 package repos
 
 import (
-	"fmt"
-	"log"
-	"os"
 	"strings"
 )
 
-func Home() string {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		log.Fatal(fmt.Errorf("error getting user home directory: %w", err))
-	}
-
-	return home
-}
-
-func ExpandHome(path string) string {
+// ExpandHome will replace ~/ with the home directory of the user.
+func ExpandHome(home, path string) string {
 	if strings.HasPrefix(path, "~/") {
-		path = Home() + path[1:]
+		path = home + path[1:]
 	}
 
 	return path
 }
 
-func ContractHome(path string) string {
-	h := Home()
-
-	if strings.HasPrefix(path, h) {
-		if len(path) == len(h) {
+// ContractHome will replace the user's home directory with a tilda(~).
+func ContractHome(home, path string) string {
+	if strings.HasPrefix(path, home) {
+		if len(path) == len(home) {
 			return "~"
 		}
 
-		path = "~" + path[len(h):]
+		path = "~" + path[len(home):]
 	}
 
 	return path